+++ Mario.h
@@ -14,10 +14,10 @@ public:
 	Mario(float startX, float startY);
 	MarioState state;
 
-	void Update(DWORD dt, vector<GameObject*>& coObjects, SceneContext* ctx) override;
+	void Update(float dt, vector<GameObject*>& coObjects, SceneContext* ctx) override;
 	void Render() override;
 	Rect GetBoundingBox() override;
-	void OnNoCollision(DWORD dt) override;
+	void OnNoCollision(float dt) override;
 	void OnCollisionWith(CollisionEvent* event) override;
 	bool IsCollidable() override { return true; }
 	bool IsBlocking() override { return true; }
